smokeview source: put frames routines into a separate file
