+++ Source/smokeview/readsmv.c
@@ -15,6 +15,9 @@
 #include "smv_endian.h"
 #include "update.h"
 #include "smokeviewvars.h"
+#ifdef pp_FRAMES
+#include "frames.h"
+#endif
 #include "IOvolsmoke.h"
 
 #ifdef pp_READBUFFER
@@ -8254,7 +8257,9 @@ typedef struct {
         nn_slice--;
         continue;
       }
-      //InitFrames(&sd->frames);
+#ifdef pp_FRAMES
+      InitFrames(&sd->frames);
+#endif
       sliceinfo_copy++;
       continue;
     }
